More excremental progress: none of this is done yet but all getting closer.
tor-browser: two new patches to change defaults and deal with the
~/.mozilla issue.

firefox-addon: ports module that factors out common code for firefox
addons.

noscript: the noscript firefox addon, which has a tbb flavor.

tbb: more work on tbb.port.mk.  Not even sure if I'm going to keep
it in the end.